Extract various searches from monolithic openLibrary client
