+++ clients/openLibraryApiClient/OpenLibraryBookResource.ts
@@ -1,0 +1,13 @@
+export type LendableItem = {
+  availability: string;
+};
+
+export type OpenLibraryBookResource = {
+  data: {
+    ebooks?: LendableItem[];
+    key: string;
+    title: string;
+  };
+  details: {};
+  isbns: string[];
+};
